delete more useless files
