+++ .env
@@ -1,0 +1,1 @@
+OPENAI_API_KEY="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"
