+++ llm_judge/judge.sh
@@ -1,4 +1,4 @@
-OPENAI_API_KEY= python -B gen_judgment.py \
+python -B gen_judgment.py \
     --bench-name "jp_bench" \
     --mode pairwise-baseline  \
     --baseline-model gpt-3.5-davinci \
