--- llm_judge/model/__init__.py
@@ -1,5 +1,0 @@
-from model.model_adapter import (
-    load_model,
-    get_conversation_template,
-    add_model_args,
-)
