--- llm_judge/model/merge.sh
@@ -1,4 +1,0 @@
-python3 -m apply_lora \
-    --base decapoda-research/llama-13b-hf \
-    --target ~/models/alpaca-lora-13b \
-    --lora ~/alpaca-lora/lora-alpaca-13b-bf16
