+++ llm_judge/model_adapter.py
@@ -1,0 +1,115 @@
+"""Model adapter registration."""
+
+import math
+import sys
+from typing import List, Optional
+import warnings
+
+if sys.version_info >= (3, 9):
+    from functools import cache
+else:
+    from functools import lru_cache as cache
+
+import accelerate
+import psutil
+import torch
+from transformers import (
+    AutoConfig,
+    AutoModel,
+    AutoModelForCausalLM,
+    AutoModelForSeq2SeqLM,
+    AutoTokenizer,
+    LlamaTokenizer,
+    LlamaForCausalLM,
+    T5Tokenizer,
+)
+
+from conversation import Conversation, get_conv_template
+
+class BaseModelAdapter:
+    """The base and the default model adapter."""
+
+    use_fast_tokenizer = True
+
+    def match(self, model_path: str):
+        return True
+
+    def load_model(self, model_path: str, from_pretrained_kwargs: dict):
+        revision = from_pretrained_kwargs.get("revision", "main")
+        try:
+            tokenizer = AutoTokenizer.from_pretrained(
+                model_path,
+                use_fast=self.use_fast_tokenizer,
+                revision=revision,
+            )
+        except TypeError:
+            tokenizer = AutoTokenizer.from_pretrained(
+                model_path,
+                use_fast=False,
+                revision=revision,
+            )
+
+        model = AutoModelForCausalLM.from_pretrained(
+            model_path, low_cpu_mem_usage=True, **from_pretrained_kwargs
+        )
+        return model, tokenizer
+
+
+    def get_default_conv_template(self, model_path: str) -> Conversation:
+        return get_conv_template("one_shot")
+
+
+# A global registry for all model adapters
+# TODO (lmzheng): make it a priority queue.
+model_adapters: List[BaseModelAdapter] = []
+
+
+def register_model_adapter(cls):
+    """Register a model adapter."""
+    model_adapters.append(cls())
+
+
+@cache
+def get_model_adapter(model_path: str) -> BaseModelAdapter:
+    """Get a model adapter for a model_path."""
+    for adapter in model_adapters:
+        if adapter.match(model_path):
+            return adapter
+    raise ValueError(f"No valid model adapter for {model_path}")
+
+
+
+def get_conversation_template(model_path: str) -> Conversation:
+    """Get the default conversation template."""
+    adapter = get_model_adapter(model_path)
+    return adapter.get_default_conv_template(model_path)
+
+
+
+
+
+
+class ChatGPTAdapter(BaseModelAdapter):
+    """The model adapter for ChatGPT"""
+
+    def match(self, model_path: str):
+        return model_path in ("gpt-3.5-turbo", "gpt-4")
+
+    def load_model(self, model_path: str, from_pretrained_kwargs: dict):
+        raise NotImplementedError()
+
+    def get_default_conv_template(self, model_path: str) -> Conversation:
+        return get_conv_template("chatgpt")
+
+
+
+
+
+# Note: the registration order matters.
+# The one registered earlier has a higher matching priority.
+
+
+register_model_adapter(ChatGPTAdapter)
+
+# After all adapters, try the default base adapter.
+register_model_adapter(BaseModelAdapter)
