+++ pyproject.toml
@@ -16,7 +16,7 @@ dependencies = [
     "accelerate", "fastapi", "gradio==3.35.2", "httpx", "markdown2[all]", "nh3", "numpy",
     "peft==0.5", "prompt_toolkit>=3.0.0", "pydantic<=2.0", "requests", "rich>=10.0.0", "sentencepiece",
     "shortuuid", "shortuuid", "tiktoken", "tokenizers>=0.12.1", "torch",
-    "transformers>=4.28.0,<4.29.0", "uvicorn", "wandb", "openai", "anthropic", "ray"
+    "transformers>=4.28.0,<4.29.0", "uvicorn", "wandb", "openai", "anthropic", "ray", "dotenv"
 ]
 
 [project.optional-dependencies]
